% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt_to_se.R
\name{load_assay_data}
\alias{load_assay_data}
\title{Load assay data based on the specified data type}
\usage{
load_assay_data(cdt, data_type)
}
\arguments{
\item{cdt}{Character. Path to the CDT Excel file.}

\item{data_type}{Character. The type of data to extract. Options are:
"peak_area", "batch_norm", "batch_norm_imputed", "mass_extracted", "log_transformed".}
}
\value{
A data frame containing the assay data.
}
\description{
Load assay data based on the specified data type
}
