% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt_to_se.R
\name{make_metadata}
\alias{make_metadata}
\title{Clean and format the metadata from the CDT file}
\usage{
make_metadata(metadata)
}
\arguments{
\item{metadata}{A data frame containing the metadata from the CDT file.}
}
\value{
A cleaned and formatted metadata data frame.
}
\description{
Clean and format the metadata from the CDT file
}
