% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt_to_se.R
\name{make_rowdata}
\alias{make_rowdata}
\title{Clean and format the row data from the CDT file}
\usage{
make_rowdata(rowdata, rowdata_key)
}
\arguments{
\item{rowdata}{A data frame containing the row data from the CDT file.}

\item{rowdata_key}{A character string specifying the column name to use as row names.}
}
\value{
A cleaned and formatted row data data frame.
}
\description{
Clean and format the row data from the CDT file
}
