% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metabolite.R
\name{plot_metabolite}
\alias{plot_metabolite}
\title{Plot Metabolite Abundance}
\usage{
plot_metabolite(metabolite, se_object, group_col, title = NULL, rm.na = TRUE)
}
\arguments{
\item{metabolite}{Character. The name of the metabolite to plot. Must be present in the rowData of \code{se_object}.}

\item{se_object}{SummarizedExperiment. The object containing metabolite data and sample metadata.}

\item{group_col}{Character (optional). The column name in colData to use for grouping samples in the plot.}

\item{title}{Character (optional). The title for the plot.}

\item{rm.na}{Logical (default TRUE). Whether to remove samples with NA values in the metabolite abundance or grouping column.}
}
\value{
A ggplot object visualizing the abundance of the specified metabolite.
}
\description{
This function generates a ggplot visualization of the abundance of a specified metabolite across samples in a SummarizedExperiment object.
}
\examples{
# plot_metabolite("00112", se_object, group_col = "Condition", title = "Abundance")

}
