% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_to_metabolon.R
\name{results_to_metabolon}
\alias{results_to_metabolon}
\title{Convert Results to Metabolon Format.}
\usage{
results_to_metabolon(
  table,
  format = "topTable",
  omics = "Transcriptomics",
  custom_colnames = NULL,
  output_file = NULL,
  save_file = TRUE
)
}
\arguments{
\item{table}{A data frame containing the results to be converted.}

\item{format}{A character string specifying the format of the input results.
Options are \code{"DESeqResults"} (default) or \code{"topTable"}.}

\item{omics}{A character string specifying the type of omics data.
Options are \code{"Transcriptomics"} (default) or \code{"Proteomics"}.}

\item{custom_colnames}{An optional character vector specifying custom column names for the output.
If \code{NULL}, default column names will be used.}

\item{output_file}{A character string specifying the path to save the output file.
If \code{NULL} and \code{save_file} is \code{TRUE}, a default filename will be generated. Needs to be csv!
@param save_file A logical value indicating whether to save the output to a file.
Default is \code{TRUE}.}
}
\value{
A data frame formatted according to the standard required by Metabolon.
}
\description{
This function converts a results table into a format compatible with Metabolon data analysis.
}
\examples{
\dontrun{
# Convert DESeqResults to Metabolon format
results_metabolon <- results_to_metabolon(
  table = deseq_results,
  format = "DESeqResults",
  omics = "Transcriptomics",
  input_features = "gene_symbol",
  save_file = TRUE
)
}
}
