% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_metabolon.R
\name{se_to_metabolon}
\alias{se_to_metabolon}
\title{Convert SummarizedExperiment to Metabolon Format}
\usage{
se_to_metabolon(
  se,
  cdt,
  input_features = "ensembl_id",
  output_file = NULL,
  organism = "Hs",
  save_file = F
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object. The input data to be converted.}

\item{cdt}{A parameter (not used in the current implementation).}

\item{input_features}{A character string specifying the type of input features.
Options are \code{"ensemble_id"} (default), \code{"gene_symbol"}, or \code{"uniprot_id"}.}

\item{output_file}{A character string specifying the path to save the output file.
If \code{NULL} and \code{save_file} is \code{TRUE}, a default filename will be generated.}

\item{organism}{A character string specifying the organism. Default is \code{"Hs"}
(Homo sapiens).}

\item{save_file}{A logical value indicating whether to save the output to a file.
Default is \code{FALSE}.}
}
\value{
A transposed assay matrix with row names and column names formatted
according to the specified \code{input_features}.
}
\description{
This function converts a \code{SummarizedExperiment} object into a format compatible
with Metabolon data analysis. It allows for customization of input features
and provides options to save the output to a file.
}
\details{
\itemize{
\item The function checks if the input object is a \code{SummarizedExperiment}.
\item Metadata is read from an external file (assumed to be in Excel format).
\item Annotations are created based on the specified organism and input features.
\item The assay matrix is transposed and formatted according to the input features.
\item If \code{save_file} is \code{TRUE}, the output is written to a CSV file.
}
}
\examples{
\dontrun{
se <- SummarizedExperiment::SummarizedExperiment()
result <- se_to_metabolon(se, cdt = NULL, input_features = "gene_symbol", save_file = TRUE)
}

}
