% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdt_to_se.R
\name{cdt_to_se}
\alias{cdt_to_se}
\title{Convert CDT File to SummarizedExperiment Object}
\usage{
cdt_to_se(
  cdt,
  output_file = NULL,
  save_file = T,
  data_type = "batch_norm_imputed"
)
}
\arguments{
\item{cdt}{Character. Path to the CDT Excel file. Must exist.}

\item{output_file}{Character. Path to save the resulting \code{SummarizedExperiment} object.
If \code{NULL} and \code{save_file} is \code{TRUE}, a default filename will be generated.}

\item{save_file}{Logical. Whether to save the resulting \code{SummarizedExperiment} object
to a file. Default is \code{TRUE}. If \code{FALSE}, the object will be returned. If a path is not provided,
a default filename will be generated based on the current date and saved in the working directory.}

\item{data_type}{Character. The type of data to extract from the CDT file. Options are:
\itemize{
\item \code{"peak_area"}: Extracts peak area data (sheet 4).
\item \code{"batch_norm"}: Extracts batch normalized data (sheet 5).
\item \code{"batch_norm_imputed"}: Extracts batch normalized and imputed data (sheet 6).
\item \code{"mass_extracted"}: Extracts mass extracted data (sheet 7).
\item \code{"log_transformed"}: Extracts log-transformed data (sheet 8).
}
Default is \code{"mbtch_norm_imputed"}.}
}
\value{
A \code{SummarizedExperiment} object containing:
\itemize{
\item \code{assays}: The assay data extracted from the CDT file.
\item \code{rowData}: Metadata for the rows (e.g., compounds).
\item \code{colData}: Metadata for the columns (e.g., samples).
}
}
\description{
This function reads a CDT (Comprehensive Data Table) Excel file and converts it into a
\code{SummarizedExperiment} object. The function allows selecting specific data types and
optionally saves the resulting object to a file.
}
\details{
The function performs the following steps:
\enumerate{
\item Validates the existence of the provided \code{cdt}.
\item Reads metadata (sheet 3) and row data (sheet 2) from the CDT file.
\item Extracts assay data based on the specified \code{data_type}.
\item Ensures consistency between metadata and assay data dimensions.
\item Constructs a \code{SummarizedExperiment} object.
\item Optionally saves the resulting object to a file if \code{save_file} is \code{TRUE}.
}
}
\examples{
\dontrun{
# Convert a CDT file to a SummarizedExperiment object
se <- cdt_to_se(
  cdt = "path/to/cdt_file.xlsx",
  save_file = FALSE,
  data_type = "log_transformed"
)
}

}
