% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_metabolon.R
\name{se_to_metabolon}
\alias{se_to_metabolon}
\title{Convert SummarizedExperiment to Metabolon Format}
\usage{
se_to_metabolon(
  se,
  cdt,
  output_file = NULL,
  save_file = TRUE,
  sample_id_column = NULL
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object containing the assay data.}

\item{cdt}{A character string specifying the path to the client data table path.}

\item{output_file}{A character string specifying the path to save the output file.
If \code{NULL} and \code{save_file} is \code{TRUE}, a default filename will be generated and it will be saved in the working directory.}

\item{save_file}{A logical value indicating whether to save the output to a file. Needs to be csv!
Default is \code{TRUE}.}

\item{sample_id_column}{A character string specifying the column name in the colData
that contains the sample IDs matching the PARENT_SAMPLE_NAME names. Default is \code{NULL}.}
}
\description{
This function converts a \code{SummarizedExperiment} object into a format compatible
with Metabolon data analysis.
}
\examples{
\dontrun{
se <- SummarizedExperiment::SummarizedExperiment()
result <- se_to_metabolon(se, cdt = "cdt.xlsx", input_features = "gene_symbol", save_file = TRUE)
}

}
