% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_to_metabolon.R
\name{se_to_metabolon}
\alias{se_to_metabolon}
\title{Convert SummarizedExperiment to Metabolon Format}
\usage{
se_to_metabolon(
  se,
  cdt,
  input_features = "ensembl_id",
  output_file = NULL,
  organism = "Hs",
  save_file = TRUE,
  sample_id_column = "CLIENT_SAMPLE_ID"
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} (or derived) object. The input data to be converted.}

\item{cdt}{The client data table from metabolon.}

\item{input_features}{A character string specifying the type of input features.
Options are \code{"ensembl_id"} (default), \code{"gene_symbol"}, or \code{"uniprot_id"}.}

\item{output_file}{A character string specifying the path to save the output file.
If \code{NULL} and \code{save_file} is \code{TRUE}, a default filename will be generated.}

\item{organism}{A character string specifying the organism. Default is \code{"Hs"}
(Homo sapiens), Options are: "Mm".}

\item{save_file}{A logical vawrlue indicating whether to save the output to a file.
Default is \code{TRUE}.}

\item{sample_id_column}{The column in the metadata that needs to be matched to the sample id in the gse.}
}
\value{
A transposed assay matrix with row names and column names formatted according to the standard required by metabolon.
}
\description{
This function converts a \code{SummarizedExperiment} object into a format compatible
with Metabolon data analysis.
}
\details{
\itemize{
\item The function checks if the input object is a \code{SummarizedExperiment}.
\item Metadata is read from the CDT.
\item Annotations are created based on the specified organism and input features.
\item The assay matrix is transposed and formatted according to the metabolon standard.
\item If \code{save_file} is \code{TRUE}, the output is written to a CSV file.
}
}
\examples{
\dontrun{
se <- SummarizedExperiment::SummarizedExperiment()
result <- se_to_metabolon(se, cdt = "cdt.xlsx", input_features = "gene_symbol", save_file = TRUE)
}

}
